refactored to component and restructured
