+++ src/constants/table.js
@@ -1,0 +1,20 @@
+export const columns = [
+  {
+    field: 'label',
+    headerName: 'Day()',
+    flex: 1,
+    minWidth: 170,
+  },
+
+  {
+    field: 'in',
+    headerName: 'In',
+    width: 120,
+  },
+]
+
+export const datFS = {
+  md: '55px',
+  sm: '30px',
+  xs: '25px',
+}
